refactor(qwen35): split config.rs into config/{model,tokenizer,tp}.rs
Split the 683-line config.rs into a config/ module with model/tokenizer/tp
submodules (k3-style mod.rs entry + config/), reducing config.rs to a re-export.
Behavior unchanged. cargo check + clippy --all-targets -D warnings + config tests
(14/14) green.
